// src/generated/api.d.ts
export interface components {
  schemas: {
// ...
    ExpenseCreateRequest: {
      /** Vendor Id */
      vendor_id?: string | null;
      /** Coa Id */
      coa_id?: string | null;
      /** Reference */
      reference?: string | null;
      /** Expense Date */
      expense_date?: string | null;
      /**
       * Amount
       * @default 0
       */
      amount: number | string;
      /**
       * Currency
       * @default NGN
       */
      currency: string;
      /** Description */
      description?: string | null;
      /**
       * Payment Status
       * @default unpaid
       */
      payment_status: string;
    };
    /** ExpenseListResponse */
    ExpenseListResponse: {
      /** Items */
      items: components['schemas']['ExpenseResponse'][];
      /** Total */
      total: number;
      /** Page */
      page: number;
      /** Page Size */
      page_size: number;
      /** Total Pages */
      total_pages: number;
      /** Has Next */
      has_next: boolean;
      /** Has Previous */
      has_previous: boolean;
    };
    /** ExpenseResponse */
    ExpenseResponse: {
      /**
       * Id
       * Format: uuid
       */
      id: string;
      /**
       * Organisation Id
       * Format: uuid
       */
      organisation_id: string;
      /** Vendor Id */
      vendor_id?: string | null;
      /** Coa Id */
      coa_id?: string | null;
      /** Reference */
      reference?: string | null;
      /**
       * Expense Date
       * Format: date
       */
      expense_date: string;
      /** Amount */
      amount: string;
      /** Currency */
      currency: string;
      /** Description */
      description?: string | null;
      /** Payment Status */
      payment_status: string;
      /**
       * Created At
       * Format: date-time
       */
      created_at: string;
      /**
       * Updated At
       * Format: date-time
       */
      updated_at: string;
      /** Created By */
      created_by?: string | null;
      /** Updated By */
      updated_by?: string | null;
      /**
       * Is Deleted
       * @default false
       */
      is_deleted: boolean;
    };
    /** ExpenseUpdateRequest */
    ExpenseUpdateRequest: {
      /** Vendor Id */
      vendor_id?: string | null;
      /** Coa Id */
      coa_id?: string | null;
      /** Reference */
      reference?: string | null;
      /** Expense Date */
      expense_date?: string | null;
      /** Amount */
      amount?: number | string | null;
      /** Currency */
      currency?: string | null;
      /** Description */
      description?: string | null;
      /** Payment Status */
      payment_status?: string | null;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}